+++ docs/examples/basic_meter/view.py
@@ -17,11 +17,7 @@ This example shows how to use the different modes to capture metrics.
 It shows the usage of the direct, bound and batch calling conventions.
 """
 from opentelemetry import metrics
-from opentelemetry.sdk.metrics import (
-    MeterProvider,
-    UpDownCounter,
-    ValueRecorder,
-)
+from opentelemetry.sdk.metrics import MeterProvider
 from opentelemetry.sdk.metrics.export import ConsoleMetricsExporter
 from opentelemetry.sdk.metrics.export.aggregate import (
     HistogramAggregator,
@@ -36,20 +32,18 @@ metrics.set_meter_provider(MeterProvider())
 meter = metrics.get_meter(__name__)
 metrics.get_meter_provider().start_pipeline(meter, ConsoleMetricsExporter(), 5)
 
-requests_counter = meter.create_metric(
+requests_counter = meter.create_counter(
     name="requests",
     description="number of requests",
     unit="1",
     value_type=int,
-    metric_type=UpDownCounter,
 )
 
-requests_size = meter.create_metric(
+requests_size = meter.create_valuerecorder(
     name="requests_size",
     description="size of requests",
     unit="1",
     value_type=int,
-    metric_type=ValueRecorder,
 )
 
 # Views are used to define an aggregation type and label keys to aggregate by