+++ opentelemetry-sdk/src/opentelemetry/sdk/trace/export/__init__.py
@@ -21,10 +21,9 @@ import typing
 from enum import Enum
 
 from opentelemetry.configuration import Configuration
-from opentelemetry.context import attach, detach, set_value
+from opentelemetry.context import Context, attach, detach, set_value
+from opentelemetry.sdk.trace import Span, SpanProcessor
 from opentelemetry.util import time_ns
-
-from .. import Span, SpanProcessor
 
 logger = logging.getLogger(__name__)
 
@@ -71,7 +70,9 @@ class SimpleExportSpanProcessor(SpanProcessor):
     def __init__(self, span_exporter: SpanExporter):
         self.span_exporter = span_exporter
 
-    def on_start(self, span: Span) -> None:
+    def on_start(
+        self, span: Span, parent_context: typing.Optional[Context] = None
+    ) -> None:
         pass
 
     def on_end(self, span: Span) -> None:
@@ -173,7 +174,9 @@ class BatchExportSpanProcessor(SpanProcessor):
         ] * self.max_export_batch_size  # type: typing.List[typing.Optional[Span]]
         self.worker_thread.start()
 
-    def on_start(self, span: Span) -> None:
+    def on_start(
+        self, span: Span, parent_context: typing.Optional[Context] = None
+    ) -> None:
         pass
 
     def on_end(self, span: Span) -> None:
@@ -210,7 +213,8 @@ class BatchExportSpanProcessor(SpanProcessor):
                     self.condition.wait(timeout)
                     flush_request = self._get_and_unset_flush_request()
                     if not self.queue:
-                        # spurious notification, let's wait again
+                        # spurious notification, let's wait again, reset timeout
+                        timeout = self.schedule_delay_millis / 1e3
                         self._notify_flush_request_finished(flush_request)
                         flush_request = None
                         continue